Refactor interface from classes; namespaces
