+++ Assets/AgoraEngine/Scripts/AgoraRtmSDK/RTMInterfaces.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b30e5db48d8c5448db8169b46f84cc39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
